+++ spec/resty/http_ng/backend/test_spec.lua
@@ -14,6 +14,7 @@ describe('test backend',function()
       local response = backend.send(req)
 
       assert.truthy(response)
+      assert.truthy(response.request)
     end)
 
     it('can verify outstanding requests', function()