+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Use `RESOLVER` before falling back to `resolv.conf` [PR #324](https://github.com/3scale/apicast/pull/324)
+- Improve error logging when failing to download configuration [PR #335](https://github.com/3scale/apicast/pull/325)
 
 ### Fixed